pyproject.toml
[project]
name = "CapstoneECommerceAgent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-opentelemetry-distro>=0.14.0",
    "strands-agents[otel]>=1.18.0",
    "bedrock-agentcore>=1.0.7",
    "bedrock-agentcore-starter-toolkit>=0.2.0",
    "boto3>=1.41.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "strands-agents>=1.18.0",
    "strands-agents-tools>=0.2.16",
    "streamlit>=1.51.0",
    "audio-recorder-streamlit>=0.0.10",
]
